BookAppBack/requirements.txt
bcrypt==3.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==9.0.1
PyJWT==2.3.0
PyMySQL==1.1.1
requests==2.25.1
